requirements.txt
fire
requests
pillow
pyinstaller